Move deduction to nu-cli/src/types/deduction
