+++ crates/nu-cli/src/commands/alias.rs
@@ -1,8 +1,8 @@
-use super::deduction::VarDeclaration;
 use crate::command_registry::CommandRegistry;
-use crate::commands::deduction::VarSyntaxShapeDeductor;
 use crate::commands::WholeStreamCommand;
 use crate::prelude::*;
+
+use crate::types::deduction::{VarDeclaration, VarSyntaxShapeDeductor};
 use deduction_to_signature::DeductionToSignature;
 use log::trace;
 use nu_data::config;
@@ -210,7 +210,7 @@ mod tests {
 
 //TODO better naming
 mod deduction_to_signature {
-    use crate::commands::deduction::{VarDeclaration, VarShapeDeduction};
+    use crate::types::deduction::{VarDeclaration, VarShapeDeduction};
     use nu_protocol::{PositionalType, Signature, SyntaxShape};
     use nu_source::Span;
